\name{WarblerG}
\docType{data}
\alias{WarblerG}
\title{Seychelles Warbler Genotypes}
\description{Genetype data collected by David Richardson from Cousin Island in 1999.  
}
\usage{WarblerG}
\format{a data frame with 307 rows and 29 columns. The first column are the unique idenitifiers for each bird, and the following columns are genotype data.  Adjacent columns beolng to the same locus.}
\source{Richardson D.S.}
\references{
Richardson \emph{et.al.} (2001) Molecular Ecology 10 2263-2273
Hadfield J.D. \emph{et al} (2006) Molecular Ecology 15 3715-31
}
\keyword{datasets}
