\name{WarblerP}
\docType{data}
\alias{WarblerP}
\title{Seychelles Warbler Phenotypes}
\description{Phenotypic data collected by David Richardson from Cousin Island in 1999.  The data are almost a complete sample of those birds that existed in the population at that time.  
}
\usage{WarblerP}
\format{a table with 307 rows and 7 columns. The columns, from left to right are: 1) a unique identifier for each bird; 2) a binary variable inbdicating whether the record belongs to an offspring; 3) the sex of each bird; 4) the territory on which the bird was recorded; 5 and 6) the latitude and longitude of that territory; 7) the behavioural status of each bird (Dominant or Subordinate)}
\source{Richardson D.S.}
\references{
Richardson \emph{et.al.} (2001) Molecular Ecology 10 2263-2273
Hadfield J.D. \emph{et al} (2006) Molecular Ecology 15 3715-31}
\keyword{datasets}
